+++ contracts/utils/Hasher.sol
@@ -12,9 +12,9 @@ import "../interfaces/IHasher.sol";
 contract Hasher is IHasher {
     /// @inheritdoc IHasher
     function packMessage(
-        uint256 srcChainSlug_,
+        uint32 srcChainSlug_,
         address srcPlug_,
-        uint256 dstChainSlug_,
+        uint32 dstChainSlug_,
         address dstPlug_,
         bytes32 msgId_,
         uint256 msgGasLimit_,
